% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sous_fonction_lasso.r
\name{coef_alpha}
\alias{coef_alpha}
\title{Lasso subfunctions}
\usage{
coef_alpha(coef, alpha)
}
\arguments{
\item{coef}{coefficient}

\item{alpha}{parameter}
}
\description{
Functions for lasso regularization
}
\details{
Contains 2 essential functions for the computation of the gradient proximent.
We call on these two functions inside the fit to calculate the coefficients
}
\examples{
\dontrun{
coef_alpha(coef,alpha)
proximent(coef,alpha)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
