% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig}
\alias{fig}
\title{Plot function}
\usage{
fig(iter, couts, cout)
}
\arguments{
\item{iter}{number of iterations}

\item{couts}{a vector with all points of cost function}

\item{cout}{new point of cost function}
}
\description{
Plot the cost function
}
\details{
Plotting the cost function every 10 iterations
}
\examples{
\dontrun{
fig(5,seq(1:5), 6)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
