% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict function}
\usage{
predict(regLog, newdata, type = "class")
}
\arguments{
\item{regLog}{Obj. Reg.Log ”is an S3 object provided by the fit (.) Function.}

\item{newdata}{are the data to be processed}

\item{type}{indicates the type of prediction {"class", "posterior"}
(predicted class or probability belonging to the classes)}
}
\description{
The predictions of the model
}
\details{
Make predictions and evaluate the performance of the model
}
\examples{
\dontrun{
data(ionosphere)
pred=predict(modele, iono_test)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
