% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_summary.r
\name{predict.summary}
\alias{predict.summary}
\title{Predict summary}
\usage{
predict.summary(pred, yreel)
}
\arguments{
\item{pred}{predition}

\item{yreel}{real distribution}
}
\description{
Display of prediction performance
}
\details{
Display of a report of prediction performance, confusion matrix
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
