% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcharges_fit.R
\name{print.Reg.Log}
\alias{print.Reg.Log}
\title{print overloaded}
\usage{
\method{print}{Reg.Log}(x, ...)
}
\arguments{
\item{x}{Obj Reg Log is an S3 object provided by the fit Function.}

\item{...}{Other parameters}
}
\description{
print overloaded
}
\details{
Generic methods print and summary overloaded
}
\examples{
\dontrun{
data(ionosphere)
model=fit(y~., data = iono_train,max_iter = 200, batch_size=100)
print(model)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
