% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_na.R
\name{process_na}
\alias{process_na}
\title{NA processing}
\usage{
process_na(x)
}
\arguments{
\item{x}{a dataframe column}
}
\description{
Replace NA from a dataframe column
}
\details{
The numeric features are replaced by the mean.
}
\examples{
\dontrun{
data(airquality)
process_na(airquality$Ozone)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
