% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sous_fonction_lasso_proximient.r
\name{proximent}
\alias{proximent}
\title{Lasso subfunctions proximent calculation}
\usage{
proximent(coef, alpha)
}
\arguments{
\item{coef}{coefficient}

\item{alpha}{parameter}
}
\description{
Functions for lasso regularization proximent calculation
}
\details{
Contains 2 essential functions for the computation of the gradient proximent.
We call on these two functions inside the fit to calculate the coefficients
}
\examples{
\dontrun{
proximent(coef,alpha)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
