% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_funct.R
\name{scale_fonct}
\alias{scale_fonct}
\title{Standardization}
\usage{
scale_fonct(x)
}
\arguments{
\item{x}{a dataframe column}
}
\description{
Standardize data
}
\details{
For each values in x, subtract x with the mean and divide by the standard deviation
}
\examples{
\dontrun{
data(airquality)
scale_fonct(airquality$Temp)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
