% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcharges_summary.R
\name{summary.Reg.Log}
\alias{summary.Reg.Log}
\title{summary overloaded}
\usage{
\method{summary}{Reg.Log}(object, ...)
}
\arguments{
\item{object}{Obj Reg Log is an S3 object provided by the fit Function.}

\item{...}{other param}
}
\description{
summary of Reg.Log overloaded
}
\details{
Generic methods print and summary overloaded
}
\examples{
\dontrun{
data(ionosphere)
model=fit(y~., data = iono_train,max_iter = 200, batch_size=100)
print(model)
summary(model)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
