% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surcharges_fit.R
\name{print.Reg.Log}
\alias{print.Reg.Log}
\title{print and summary}
\usage{
\method{print}{Reg.Log}(object)
}
\arguments{
\item{object}{Obj Reg Log is an S3 object provided by the fit Function.}
}
\description{
print and summary overloaded
}
\details{
Generic methods print and summary overloaded
}
\examples{
\dontrun{
data(ionosphere)
model=fit(y~., data = iono_train,max_iter = 200, batch_size=100)
print(model)
summary(model)
}
}
\author{
Sami Ait Tilat, Afaf Ben Haj, Marie Vachet
}
